requirements/base.txt

# pip3 install -r requirements/base.txt 

# to read .env
python-dotenv

fastapi[all]


sqlalchemy
alembic

psycopg2
psycopg2-binary

# for async orm
asyncpg

celery

fastapi-sqlalchemy

retry

Jinja2

# 
python-jose[cryptography]
passlib
cryptography

jwt


python-multipart


requests

